Impl UserService

+++ src/main/java/entity/User.java
@@ -2,17 +2,17 @@ package entity;
 
 public class User {
 
-    private int id;
+    private long id;
     private String login;
     private String password;
 
-    public User(int id, String login, String password) {
+    public User(long id, String login, String password) {
         this.id = id;
         this.login = login;
         this.password = password;
     }
 
-    public int getId() {
+    public long getId() {
         return id;
     }
